Add transitions between frames
Adds clip property to Frame so that it is easier to go from a
Transition source/target frame back to the clip containing it.

Adds ``next`` property to Frame for the next frame in the clip. This
makes it easier to determine all the possible frames that can be
displayed next by combining the next property and all transitions from
the frame.

Towards #6.