Add properties JSON to frame table
This allows for possibly-sparse properties to be stored on the frame
objects.

The first example of this is an originalImage property, which can be
used to reconstruct how the Pikov file was created from a spritesheet.
This is a step towards #20 because it can also be used to create a "bag
of files" version of a SQLite Pikov file by using the original
spritesheet as part of the bag of files.