Use user-provide string IDs for clips instead of integers
I was using integer IDs before because that's what I did in the PICO-8
prototype, but with string IDs I don't have to include the clips that
I'm not using (walk, strut, jump, turn) just to make the numbers line
up with the PICO-8 version.

Temporarily disables the mirror version of clips until after I get a
preview rendering of the Pikov file working.

Closes #18