Remove clip table in favor of frame with transitions.
* Remove Frame.next in preparation for using transitions everywhere

* Refactor Clip class(s).

A Clip is now more like what MultiClip was. It's a handy way of editing
a Pikov animation but no longer a core concept. A Frame can belong to 0
or more clips.

* Fix clip GIF preview for unchanging image.

When the image remains unchanged between frames, increase the duration
rather than attempting to add a frame. If only a single image is present
in the preview, use an integer for duration.

Bump the version of Pillow for
https://github.com/python-pillow/Pillow/issues/3073